refactor: 하이라이트의 여러줄에 대한 도메인 분리
